@@ -5,15 +5,14 @@
         }
     },
     "require": {
-        "php": ">=5.4.0",
-        "twig/twig": "1.12.*",
-        "slim/slim": "2.*",
-        "slim/extras": "2.*",
-        "slim/views": "0.1.*",
+        "php": ">=7.2.0",
+        "slim/slim": "4.*",
         "michelf/php-markdown": "1.4.1",
-        "mnapoli/front-yaml": "~1.2"
+        "mnapoli/front-yaml": "~1.2",
+        "slim/psr7": "^1.0",
+        "slim/twig-view": "3.0.0",
+        "php-di/php-di": "^6.0",
+        "ext-json": "*"
     },
-    "require-dev": {
-        "phpunit/phpunit": "~4.0"
-    }
+    "require-dev": {}
 }
